# openapi.yaml
openapi: 3.1.0
info:
  title: Gates API
  version: 0.0.0
  description: OpenAPI specification for the gates API
  license:
    name: MIT Licence
    identifier: MIT
    url: https://github.com/consid-germany/gates/blob/main/LICENSE
  contact:
    url: https://github.com/consid-germany/gates
...
servers:
  - url: http://localhost:9000/api
  - url: https://gates.consid.tech/api
security: []
paths:
...
  /gates/{group}/{service}/{environment}/state:
    get:
      operationId: get_gate_state
      tags:
        - gate
      summary: Get the state of a gate by group, service and environment
      description: This should be used if you want to explicitly know the state of a gate.
      parameters:
        - $ref: '#/components/parameters/group'
        - $ref: '#/components/parameters/service'
        - $ref: '#/components/parameters/environment'
      responses:
        '200':
          description: The state of the requested gate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateStateRep'
        '204':
          $ref: '#/components/responses/NoContent'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: update_gate_state
      tags:
        - gate
      summary: Update the state of a gate
      description: With this operation you can change the `state` of the gate, e.g. switch the gate state from `closed` to `open` and vice versa.
      parameters:
        - $ref: '#/components/parameters/group'
        - $ref: '#/components/parameters/service'
        - $ref: '#/components/parameters/environment'
      requestBody:
        $ref: '#/components/requestBodies/UpdateGateStatePayload'
      responses:
        '200':
          description: Returns the updated gate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /gates/{group}/{service}/{environment}/display-order:
    put:
      operationId: update_display_order
      tags:
        - gate
      summary: Update the display order of a gate
      description: This can be used to sort the frontend representation of the gates. Sorting has to be done by the frontend.
      parameters:
        - $ref: '#/components/parameters/group'
        - $ref: '#/components/parameters/service'
        - $ref: '#/components/parameters/environment'
      requestBody:
        $ref: '#/components/requestBodies/UpdateDisplayOrderPayload'
      responses:
        '200':
          description: Returns the updated gate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /gates/{group}/{service}/{environment}/comments:
    post:
      operationId: add_comment
      tags:
        - gate
      summary: Add a comment to a gate
      description: | 
        You can add additional notes as `comment` to a gate. E.g. Providing information why you changed the gate state.
        You may use the autogenerated id of the comment to specifically remove it using `delete comment`.
      parameters:
        - $ref: '#/components/parameters/group'
        - $ref: '#/components/parameters/service'
        - $ref: '#/components/parameters/environment'
      requestBody:
        $ref: '#/components/requestBodies/AddCommentPayload'
      responses:
        '200':
          description: Returns the updated gate.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
          links:
             DeleteComment:
               operationId: delete_comment
               parameters:
                 group: '$response.body#/group'
                 service: '$response.body#/service'
                 environment: '$response.body#/environment'
                 comment_id: '$response.body#/comments/0/id'
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    group:
      name: group
      in: path
      description: Group name of the gate
      required: true
      schema:
        type: string
        pattern: '^[0-9a-zA-Z-_]+$'
        minLength: 1
    service:
      name: service
      in: path
      description: Service name of the gate
      required: true
      schema:
        type: string
        pattern: '^[0-9a-zA-Z-_]+$'
        minLength: 1
    environment:
      name: environment
      in: path
      description: Environment name of the gate
      required: true
      schema:
        type: string
        pattern: '^[0-9a-zA-Z-_]+$'
        minLength: 1
...
  requestBodies:
...
    UpdateGateStatePayload:
      description: This is needed to update the state of the gate.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              state:
                $ref: "#/components/schemas/GateState"
            required:
              - state
    UpdateDisplayOrderPayload:
      description: This parameter is needed to change the display order.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              display_order:
                type: number
                description: Is expected to be a positive number, but can basically be anything.
                example: 1
            required:
              - display_order
    AddCommentPayload:
      description: If you want to add a comment, you need to set this.
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                minLength: 1
                description: This should be a thoughtful description of why this gate is here.
                example: Running some tests (Max)
            required:
              - message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: string
            example: group, service and environment must not be empty
    Forbidden:
      description: Unauthorized request blocked.
    Conflict:
      description: Request could not be processed because of a conflict in the current state of the resource.
    NoContent:
      description: Requested resource could not be found.
    UnprocessableContent:
      description: Request body could not be processed.
    InternalServerError:
      description: Request failed (internal server error).
      content:
        application/json:
          schema:
            type: string
            example: 'service error: ResourceNotFoundException: Cannot do operations on a non-existent table'
  schemas:
...
    Gate:
      type: object
      description: The gate is lowest but most important part of the hierachy. This defines the actual object, that will be altered by all actions. You can open and close it.
      properties:
        group:
          type: string
          minLength: 1
          description: The group that surrounds the service.
          example: some-gate-group
        service:
          type: string
          minLength: 1
          description: The service that contains the environment.
          example: some-service
        environment:
          type: string
          minLength: 1
          description: The environment for the specific gate.
          example: develop
        state:
          $ref: "#/components/schemas/GateState"
        display_order:
          type: number
          description: The way to describe how you want to arrange your gates.
          example: 1
        comments:
          type: array
          description: The comment describe why you open or closed the gate.
          items:
            $ref: "#/components/schemas/Comment"
        last_updated:
          type: string
          description: Changes when a comment or gate is changed.
          format: date-time
          example: 2023-05-26T21:36:18.345195Z
      required:
        - group
        - service
        - environment
        - state
        - comments
        - last_updated
    GateStateRep:
      type: object
      description: Wrapper object for the state of a gate.
      properties:
        state:
          $ref: "#/components/schemas/GateState"
      required:
        - group
    GateState:
      type: string
      description: This can only be open or closed.
      example: closed
      enum:
        - open
        - closed
    Comment:
      type: object
      description: Comments will help you identify the reasons for the gate.
      properties:
        id:
          type: string
          minLength: 1
          description: Identifies the comment.
          example: e9qo5ibvyvk8ffmhu6qe6p30
        message:
          type: string
          minLength: 1
          description: Very important text to show why you did it.
          example: Running some tests (Max)
        created:
          type: string
          description: The exact moment the comment was created.
          format: date-time
          example: 2023-05-26T21:36:18.345195Z
      required:
        - id
        - message
        - created
...